fix: properly type more OpenApi models

+++ libs/api/transaction/data-access/src/lib/entities/minimum-rent-exemption-balance-response.entity.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger'
 
 export class MinimumRentExemptionBalanceResponse {
-  @ApiProperty({ type: 'float' })
+  @ApiProperty({ type: 'integer' })
   lamports: number
 }
